fixing several bugs in and augmenting of the benchmarks; run results for the local dev machine are copied at the top of the source files.
Conclusion: Win32 API is *nearly* the fastest (which makes me wonder what tricks the MSVC runtime library is pulling) as the C FILE API (fopen/fread/fclose), in BINARY mode, using TextBuffer instead of std::string, is the very fastest kid on the block when it comes to getting file content into a memory buffer, so that we can traverse it with a std::string_view.

The question now is: can we make that line splitting code any faster?